container/requirements.txt
google-cloud-storage
pandas
google-cloud-pubsub
fastapi
uvicorn